- add abstract controller test; - refactor log messages; - refactor and fix restaurant controller; - add restaurant tests; - redefine equals and hashcode for restaurant DTO; - fix restaurant test data; - fix restaurant util methods; - refactor security config; - fix user test data; - refactor vote service; - add vote service tests; - fix vote test data.

+++ src/main/java/by/bestlunch/service/VoteService.java
@@ -6,6 +6,7 @@ import by.bestlunch.persistence.repository.MenuRepository;
 import by.bestlunch.persistence.repository.RestaurantRepository;
 import by.bestlunch.persistence.repository.UserRepository;
 import by.bestlunch.persistence.repository.VoteRepository;
+import by.bestlunch.util.DateTimeUtil;
 import by.bestlunch.util.exception.MenuException;
 import by.bestlunch.util.exception.VoteException;
 import by.bestlunch.web.dto.VoteDto;
@@ -15,7 +16,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -46,30 +46,41 @@ public class VoteService {
     }
 
     @Transactional
-    public void createOrUpdate(int userId, int restaurantId) {
-        Vote newVote = new Vote(userRepository.get(userId), restaurantRepository.get(restaurantId), LocalDate.now());
+    public Vote createOrUpdate(int userId, int restaurantId) {
+        Vote newVote = new Vote(userRepository.get(userId), restaurantRepository.get(restaurantId), DateTimeUtil.now());
 
-        VoteDto todayVote = voteRepository.getUserVote(userId, LocalDate.now())
+        VoteDto todayVote = voteRepository.getUserVote(userId, DateTimeUtil.now())
                 .map(v -> {
                     v.setRestaurant(restaurantRepository.get(restaurantId));
                     return new VoteDto(v, false);
                 })
                 .orElseGet(() -> new VoteDto(newVote, true));
 
-        if (LocalTime.now().isAfter(DEFAULT_EXPIRED_TIME)) {
+        if (DateTimeUtil.nowTime().compareTo(DEFAULT_EXPIRED_TIME) > 0) {
             throw new VoteException("");
         }
 
-        List<Menu> dailyMenu = menuRepository.getAllByDate(restaurantId, LocalDate.now());
+        List<Menu> dailyMenu = menuRepository.getAllByDate(restaurantId, DateTimeUtil.now());
         if (dailyMenu.size() == 0) {
             throw new MenuException("");
         }
 
-        voteRepository.save(todayVote.getVote());
+        return voteRepository.save(todayVote.getVote());
     }
 
     public void delete(int restaurantId, int userId) {
-        Optional<Vote> userVote = voteRepository.getUserVote(userId, LocalDate.now());
+        if (DateTimeUtil.nowTime().compareTo(DEFAULT_EXPIRED_TIME) > 0) {
+            throw new VoteException("");
+        }
+
+        Optional<Vote> userVote = voteRepository.getUserVote(userId, DateTimeUtil.now());
         userVote.ifPresent(vote -> voteRepository.delete(userVote.get().id(), restaurantId, userId));
+    }
+
+    public Vote get(int userId, LocalDate localDate) {
+        Optional<Vote> userVote = voteRepository.getUserVote(userId, localDate);
+        return userVote.map(vote ->
+                new Vote(vote.id(),
+                        vote.getUser(), vote.getRestaurant(), vote.getVotingDate())).orElseThrow();
     }
 }
